test(metrics): deduplicate TestTensorToPython tests
Extract common tensor_to_python tests into shared test_utils.py module to eliminate duplication across classification and detection metric tests.

Made-with: Cursor
